refactor: 테스트 패키지 수정
