+++ app/src/test/java/woowacourse/shopping/fixture/FakeCartRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.fixture
+package woowacourse.shopping.fixture
 
 import woowacourse.shopping.domain.CartRepository
 import woowacourse.shopping.model.Product
