+++ app/src/test/java/woowacourse/shopping/fixture/FakeProductRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.fixture
+package woowacourse.shopping.fixture
 
 import woowacourse.shopping.domain.ProductRepository
 import woowacourse.shopping.model.Product
